add version 0.3.0
