+++ setup.py
@@ -1,7 +1,8 @@
-from setuptools import setup, find_packages
+from setuptools import setup
+from setuptools import find_packages
 
 # load the README file.
-with open("README.md", "r") as fh:
+with open(file="README.md", mode="r") as fh:
     long_description = fh.read()
 
 setup(
@@ -16,7 +17,7 @@ setup(
 
     # I'm in alpha development still, so a compliant version number is a1.
     # read this as MAJOR VERSION 0, MINOR VERSION 1, MAINTENANCE VERSION 0
-    version='0.2.9',
+    version='0.3.0',
 
     # here is a simple description of the library, this will appear when
     # someone searches for the library on https://pypi.org/search
