+++ td/exceptions.py
@@ -1,14 +1,16 @@
 class TknExpError(Exception):
     """Raise exception when refresh or access token is expired.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -16,14 +18,16 @@ class TknExpError(Exception):
 class ExdLmtError(Exception):
     """Raise exception when exceeding query limit of the server.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -31,14 +35,16 @@ class ExdLmtError(Exception):
 class NotNulError(Exception):
     """Raise exception when a null value is passed into non-null field.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -47,14 +53,16 @@ class ForbidError(Exception):
     """Raise forbidden exception. This usually occurs when the app does
     not have access to the account.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -62,14 +70,16 @@ class ForbidError(Exception):
 class NotFndError(Exception):
     """Raise exception when criteria is not found.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -78,14 +88,16 @@ class ServerError(Exception):
     """Raise exception when there is an error with the service or the server
     cannot provide response.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
@@ -94,14 +106,16 @@ class GeneralError(Exception):
     """Raise exception for all other status code >400 errors which are not
     defined above.
 
-    Args:
-        Exception (Exception): The base python exception class
+    Arguments:
+    ----
+    Exception (Exception): The base python exception class
     """
     def __init__(self, message):
         """Print out message for this exception.
 
-        Args:
-            message (str): Pass in the message returned by the server.
+        Arguments:
+        ----
+        message (str): Pass in the message returned by the server.
         """
         self.message = message
         super().__init__(self.message)
