+++ tests/unit/test_order.py
@@ -1,16 +1,14 @@
 import unittest
-
 import td.enums as td_enums
 
 from unittest import TestCase
-from datetime import datetime
-from datetime import timedelta
 from configparser import ConfigParser
 
 from td.orders import Order
 from td.orders import OrderLeg
 from td.client import TDClient
 from td.stream import TDStreamerClient
+
 
 class TDSession(TestCase):
 
@@ -30,10 +28,10 @@ class TDSession(TestCase):
 
         # Initalize the session.
         self.td_session = TDClient(
-            client_id=CLIENT_ID, 
-            redirect_uri=REDIRECT_URI, 
+            client_id=CLIENT_ID,
+            redirect_uri=REDIRECT_URI,
             credentials_path=JSON_PATH,
-            account_number = ACCOUNT_NUMBER
+            account_number=ACCOUNT_NUMBER
         )
 
         self.td_order = Order()
@@ -49,15 +47,41 @@ class TDSession(TestCase):
     def test_define_simple_order(self):
         """Test creating a simple order."""
 
-        self.td_order.order_session(session=td_enums.ORDER_SESSION.NORMAL)
-        self.td_order.order_duration(duration=td_enums.DURATION.GOOD_TILL_CANCEL)
-        
-        self.td_order_leg.order_leg_instruction(instruction=td_enums.ORDER_INSTRUCTIONS.SELL)
-        self.td_order_leg.order_leg_price(price=112.50)
-        self.td_order_leg.order_leg_quantity(quantity=10)
-        self.td_order_leg.order_leg_asset(asset_type=td_enums.ORDER_ASSET_TYPE.EQUITY, symbol='MSFT')
+        # Add the Order session.
+        self.td_order.order_session(
+            session=td_enums.ORDER_SESSION.NORMAL
+        )
 
-        self.td_order.add_order_leg(order_leg=self.td_order_leg)
+        # Add the Order duration.
+        self.td_order.order_duration(
+            duration=td_enums.DURATION.GOOD_TILL_CANCEL
+        )
+
+        # Add the Order Leg Instruction.
+        self.td_order_leg.order_leg_instruction(
+            instruction=td_enums.ORDER_INSTRUCTIONS.SELL
+        )
+
+        # Add the Order Leg price.
+        self.td_order_leg.order_leg_price(
+            price=112.50
+        )
+
+        # Add the Order Leg quantity.
+        self.td_order_leg.order_leg_quantity(
+            quantity=10
+        )
+
+        # Add the Order Leg Asset.
+        self.td_order_leg.order_leg_asset(
+            asset_type=td_enums.ORDER_ASSET_TYPE.EQUITY,
+            symbol='MSFT'
+        )
+
+        # Add the Order Leg.
+        self.td_order.add_order_leg(
+            order_leg=self.td_order_leg
+        )
 
         correct_dict = {
             "session": "NORMAL",
@@ -68,7 +92,7 @@ class TDSession(TestCase):
                     "price": 112.5, "quantity": 10,
                     "instrument": {
                         "assetType":
-                        "EQUITY", 
+                        "EQUITY",
                         "symbol": "MSFT"
                     }
                 }
@@ -83,6 +107,7 @@ class TDSession(TestCase):
         self.td_session = None
         self.td_order = None
         self.td_order_leg = None
+
 
 if __name__ == '__main__':
     unittest.main()
